fix bugs in ContainerManager.Animation and in EditText

+++ MicroUI/src/microui/component/EditText.java
@@ -110,7 +110,8 @@ public class EditText extends Component implements Scrollable, KeyPressable {
 		cursor.draw(graphics);
 		graphics.endDraw();
 
-		ctx.image(graphics, getX(), getY(), getWidth(), getHeight());
+		ctx.image(graphics, (int) getX(), (int) getY(), (int) getWidth(), (int) getHeight());
+	
 		scrollV.draw();
 		scrollH.draw();
